% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rhoedown-package.R
\docType{package}
\name{rhoedown}
\alias{rhoedown}
\alias{rhoedown-package}
\title{Markdown rendering for R}
\description{
\pkg{Rhoedown} is a plain-text formatting syntax that can be converted to
XHTML or other formats. This package provides R bindings to the Hoedown
(\url{https://github.com/hoedown/hoedown}) markdown rendering library.
}
\details{
The R function \code{\link{markdownToHTML}} renders a markdown file to HTML
(respecting the specified \code{\link{markdownExtensions}} and
\code{\link{markdownHTMLOptions}}).

The package also exports the underlying Hoedown C extension API which enables
creating and calling custom renderers using the \code{\link{renderMarkdown}}
function.

To learn more about markdown syntax see:

\url{http://en.wikipedia.org/wiki/Markdown}
}
\author{
Jonathan Gilligna, JJ Allaire, Jeffrey Horner, Vicent Marti, and Natacha Porte

  Maintainer: Jonathan Gilligan  <jonathan.gilligan@gmail.com>
}
\seealso{
\code{\link{markdownToHTML}} \code{\link{renderMarkdown}}
}
\keyword{package}

